Updated style and added alert, when you collide with an object
I've updated the theme, and changed it away from the dino and the bird, and made it a heart and a bone.
I've also added an alert which triggers, when the heart hits the bone.
My current issue is that I have to reload the site to restart the game, and it doesn't do so on it's own