Added comments to js,  Made it so the animations doesn't run, and started changing code
Since I don't want the same animations on the bones and the jump on the heart, I've started considering another way of making my character move. Currently watching a tutorial.

Also I added comments in JS to define what I want to do with the code, so that it is easier to have an overview of what's being done